// src/enum.ts
export enum Rank { one = 1, two = 2, three = 3, four = 4, five = 5, six = 6, seven = 7, eight = 8 }
export enum File {
    a = 'a',
    b = 'b',
    c = 'c',
    d = 'd',
    e = 'e',
    f = 'f',
    g = 'g',
    h = 'h'
}
export enum Team { black = 'black', white = 'white' }
export enum PieceDescription {
    rook = 'rook',
    pawn = 'pawn',
    knight = 'knight',
    bishop = 'bishop',
    king = 'king',
    queen = 'queen'
}